Split test into multiple files
